% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pricing_package.R
\docType{package}
\name{paws.pricing-package}
\alias{paws.pricing}
\alias{paws.pricing-package}
\title{paws.pricing: AWS Price List Service}
\description{
AWS Price List Service API (AWS Price List Service)
is a centralized and convenient way to programmatically query Amazon
Web Services for services, products, and pricing information. The AWS
Price List Service uses standardized product attributes such as
`Location`, `Storage Class`, and `Operating System`, and provides
prices at the SKU level. You can use the AWS Price List Service to
build cost control and scenario planning tools, reconcile billing
data, forecast future spend for budgeting purposes, and provide cost
benefit analysis that compare your internal workloads with AWS.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
